% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HClus_dendrograms.R
\name{BHclus}
\alias{BHclus}
\title{fBHclus}
\usage{
BHclus(
  df_m,
  meth = "ward.D2",
  dist = "euclidean",
  varCat1,
  value,
  nb = 100,
  fig = TRUE,
  ni = 1,
  nf = 10
)
}
\arguments{
\item{ni, nf:}{first and last columns corresponding to categorical variables  (default values, ni=1, nf=10)}

\item{df_m:}{dataframe containing peaks and metadata}

\item{meth:}{hierarchical clustering algorithm ("ward2", default value), other values: "average", "ward.D", "single", "complete", "mcquitty", "median" or "centroid"}

\item{dist:}{distance ("euclidean", default value), "maximum", "manhattan", "canberra", "binary" "correlation", "uncentered", "abscor"}

\item{varCat1:}{categorical variable, example: "Genre","Taxonomie", " Nutrition", "Ruche"...}

\item{value:}{value of catVar1 "Lactobacillus" ("Genre"), Taxonomie("Pediococcus pentosaceus"), "Erica cinerea" ("Nutrition"),...}

\item{nb:}{n of bootstrap iterations (nb=100, default)}

\item{fig:}{boolean variable to indicate output figure (fig=TRUE, default value)}
}
\value{
output cluster and figure
}
\description{
function wrapper of pvclust for building dendrograms labeled with bootstrap probabilities for isolates chosen according to varCat1 levels
}
\examples{
    dft<-BHclus(df_Peaks, varCat1="Genre", value="All", nb=500),
              dft<-BHclus(df_Peaks, varCat1="Genre", value="Lactobacillus", nb=500)
              dft<-BHclus(df_Peaks, varCat1="Taxonomie", value="Pediococcus pentosaceus"),
              dft<-BHclus(df_Peaks, meth="complete", dist="canberra",varCat1="Taxonomie", value="Pediococcus pentosaceus")
      source: https://academic.oup.com/bioinformatics/article/22/12/1540/207339
              https://www.rdocumentation.org/packages/pvclust/versions/2.2-0/topics/pvclust
              https://www.rdocumentation.org/packages/stats/versions/3.2.1/topics/dist
}
