% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DProcessing.R
\name{CopyRenameXml}
\alias{CopyRenameXml}
\title{CopyRenameXml}
\usage{
CopyRenameXml(Dirp, newdir)
}
\arguments{
\item{Dirp:}{folder containing the mzXML files,}

\item{newdir:}{the folder to which the mzXML files are copied}
}
\description{
CopyRenameXml is a function aimed at copying and renaming the mzXML files not being in the folders: "BTS", "BTS_Validation", "CTL", "Autocalibration"
}
\examples{
 CopyRenameXml("Bees project 2019-2020", "newdata")
}
