% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HClus_dendrograms.R
\name{Dendrogram_pairComp}
\alias{Dendrogram_pairComp}
\title{Dendrogram_pairComp}
\usage{
Dendrogram_pairComp(
  df_m,
  varCat1,
  value,
  methCorr = "cophenetic",
  meth1 = "ward.D2",
  meth2 = "mcquitty",
  meth3 = "complete",
  meth4 = "centroid",
  meth5 = "single",
  meth6 = "average",
  dist1 = "euclidean",
  dist2 = "euclidean",
  dist3 = "euclidean",
  dist4 = "euclidean",
  dist5 = "euclidean",
  dist6 = "euclidean",
  graph = "Complex",
  nc = 3
)
}
\arguments{
\item{meth1, meth2...meth6:}{hierarchical clustering algorithms "ward2" ,"average", "ward.D", "single", "complete", "mcquitty", "median" or "centroid"}

\item{dist1, dist2, ...dist6:}{distances "euclidean", "maximum", "manhattan", "canberra", "binary" "minkowski"}

\item{df_m:}{dataframe containing peaks and metadata}

\item{methCorr:}{dendrogram correlation method ("cophenetic", default value), others: "baker", "common_nodes", "FM_index"}

\item{varCat1:}{categorical variable for choosing isolates, examples: "Taxonomie"    ,"Genre",    "Date.d.analyse"    ,"Origine","Ruche", "Nutrition"    , "Date.de.récolte"    , "Lieu.de.la.ruche"}

\item{value:}{level of catVar1, examples: "Lactobacillus" ("Genre"), Taxonomie("Pediococcus pentosaceus"), "Erica cinerea" ("Nutrition"),...}
}
\value{
untangled and tangled dendrograms, dendrogram correlation matrix and statistics
 @examples    mc<-Dendrogram_pairComp(df_Peaks, varCat1="Taxonomie", value="Pediococcus pentosaceus"),
              mc<-Dendrogram_pairComp(df_Peaks, varCat1="Nutrition", value="Marrubium vulgare")
              source: https://cran.r-project.org/web/packages/dendextend/vignettes/dendextend.html#:~:text=The%20dendextend%20package%20offers%20a,its%20branches%2C%20nodes%20and%20labels.
                     https://www.rdocumentation.org/packages/dendextend/versions/1.14.0
                     https://www.r-graph-gallery.com/340-custom-your-dendrogram-with-dendextend.html
                     https://academic.oup.com/bioinformatics/article/31/22/3718/240978
                     https://cran.rstudio.com/web/packages/dendextend/vignettes/Cluster_Analysis.html
                     https://www.datanovia.com/en/lessons/comparing-cluster-dendrograms-in-r/
                     https://rdrr.io/cran/dendextend/man/untangle.html
}
\description{
Dendrogram_pairComp makes pairwise dendrogram comparisons and correlations
}
