% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDProcessing.R
\name{DropCMetadata}
\alias{DropCMetadata}
\title{DropCMetadata}
\usage{
DropCMetadata(dm, fileout, Co)
}
\arguments{
\item{dm:}{input dataframe,}

\item{fileout:}{csv file for both joined dataframes}
}
\value{
output dataframe and csv file
}
\description{
removes variables from the metadata csv file
}
\examples{
dfc<-DropCMetadata(dfm,"DCFilered_Meta.csv", c("Date d'analyse","Date de récolte"))
}
