% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaldiTOFSpectraPA.R
\docType{package}
\name{MaldiTOFSpectra}
\alias{MaldiTOFSpectra}
\title{MaldiTOFSpectraPA package}
\description{
Import and processing of MALDI_TOF spectra and metadata. Clustering of
MALDI_TOF spectra (significant peaks)and cluster validation.
Dendrograms, Correlograms, PCA and CMS.
}
\author{
Miguel Sautié Castellanos  \email{miguel.sautie.castellanos@umontreal.ca}
}
