% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DClus_MS.R
\name{MDS_Clus}
\alias{MDS_Clus}
\title{MDS_Clus,}
\usage{
MDS_Clus(
  df_m,
  varCat1,
  value,
  dist = "euclidean",
  nc = 3,
  grah = "lab_mdsclus"
)
}
\arguments{
\item{df_m:}{dataframe containing peaks and metadata}

\item{dist:}{distances dist: "euclidean" (default value), "maximum", "manhattan", "canberra", "binary" or "minkowski"}

\item{varCat1:}{categorical variable for choosing isolates, examples: "Taxonomie"    ,"Genre",    "Date.d.analyse"    ,"Origine","Ruche", "Nutrition"    , "Date.de.récolte"    , "Lieu.de.la.ruche"}

\item{value:}{level of the chosen categorical variable catVar1, examples: "Lactobacillus" ("Genre"), Taxonomie("Pediococcus pentosaceus"), "Erica cinerea" ("Nutrition"),...}

\item{grah:}{graphs: lab_mdsclus,(default value) lb_mds, mdsclaing}
}
\value{
figures and statistics
}
\description{
function for Multidimensional scaling and kmeans-based analysis of Maldi_Tof spectra
}
\examples{
   g<-MDS_Clus(df_Peaks, varCat1="Genre", value="Lactobacillus")
     source: https://rstudio-pubs-static.s3.amazonaws.com/274936_050c742fb3514bbaa87ce6ee2686af8c.html
              http://www.sthda.com/english/articles/31-principal-component-methods-in-r-practical-guide/122-multidimensional-scaling-essentials-algorithms-and-r-code/
              http://ugrad.stat.ubc.ca/R/library/mva/html/cmdscale.html
}
