% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDProcessing.R
\name{MergeMetaData}
\alias{MergeMetaData}
\title{MergeMetaData}
\usage{
MergeMetaData(df, Dp, filein, fileout, keyCode = "Maldi_code")
}
\arguments{
\item{df:}{input dataframe (output of PMetaData1), Dp: folder containing the metadata files,}

\item{filein:}{xlsx files for metadata}

\item{fileout:}{csv file for both joined dataframes,}

\item{keyCode:}{key code used for Dataframe merging, "Maldi_code" (Default)}
}
\value{
both dataframes joined in a new dataframe and an output csv file
}
\description{
Merges two dataframes, one is the output of the funct. PMetaData1 and the other comes from a xlsx file. Merging is done through the Maldi code. Therefore it is important to ensure error-free transcription of Maldi codes.
}
\examples{
dfm<-MergeMetaData(Rn, "Bees project 2019-2020","Bees metadata.xlsx", "Filtered_Meta.csv")
}
