% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_PCA_Clus.R
\name{PCA_Clus}
\alias{PCA_Clus}
\title{PCA_Clus}
\usage{
PCA_Clus(
  df_m,
  varCat1,
  value,
  meth = "ward",
  dist = "euclidean",
  graph = "factorMapClus",
  pc = 3,
  ni = 1,
  nf = 10
)
}
\arguments{
\item{ni, nf:}{first and last columns corresponding to categorical variables  (default values, ni=1, nf=10)}

\item{df_m:}{dataframe containing peaks and metadata}

\item{dist:}{distances, "euclidean", (default value),  "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski", "pearson", "spearman" or "kendall".}

\item{varCat1:}{categorical variable for choosing isolates, examples: "Taxonomie"    ,"Genre",    "Date.d.analyse"    ,"Origine","Ruche", "Nutrition"    , "Date.de.récolte"    , "Lieu.de.la.ruche"}

\item{value:}{level of catVar1, examples: "Lactobacillus" ("Genre"), Taxonomie("Pediococcus pentosaceus"), "Erica cinerea" ("Nutrition"),...}

\item{meth:}{clustering method, ward(default), "average" ,"single","complete‘}

\item{graph:}{visual analysis, "dendf", "dendh", (dendograms) "factorMapf", "factorMaph", "factorMapClus",(factor maps) (default value, graph="factorMapClus" )}

\item{pc:}{number of principal components (pc=3, default value)}
}
\value{
figures and statistics
}
\description{
Principal Component Analysis and clustering of Maldi_Tof spectra
}
\examples{
 pc<-PCA_Clus(df_Peaks, varCat1="Genre", value="Lactobacillus"),
           pc<-PCA_Clus(df_Peaks, varCat1="Genre", value="Lactobacillus", graph="dendh")
           pc<-PCA_Clus(df_Peaks, varCat1="Genre", value="Lactobacillus", graph="dendf"),
           pc<-PCA_Clus(df_Peaks, varCat1="Genre", value="Lactobacillus", graph="factorMapf")
    source: https://www.rdocumentation.org/packages/FactoMineR/versions/2.2/topics/PCA
           https://www.rdocumentation.org/packages/FactoMineR/versions/2.2/topics/HCPC
           http://www.sthda.com/english/articles/31-principal-component-methods-in-r-practical-guide/117-hcpc-hierarchical-clustering-on-principal-components-essentials/
           https://rpkgs.datanovia.com/factoextra/
           http://factominer.free.fr/factomethods/hierarchical-clustering-on-principal-components.html
}
