% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HClus_dendrograms.R
\name{Phyclus}
\alias{Phyclus}
\title{Phyclus}
\usage{
Phyclus(
  df_m,
  meth = "ward.D2",
  d = "euclidean",
  varCat1,
  value,
  dendrogram = "phylogram",
  nc = 4,
  ni = 1,
  nf = 10
)
}
\arguments{
\item{ni, nf:}{first and last columns corresponding to categorical variables  (default values, ni=1, nf=10)}

\item{df_m:}{dataframe containing peaks and metadata}

\item{meth:}{hierarchical clustering algorithm ("ward2", default value), other values: "average", "ward.D", "single", "complete", "mcquitty", "median" or "centroid"}

\item{dist:}{distance ("euclidean", defaul), "euclidean", "maximum", "manhattan", "canberra", "binary" "minkowski"}

\item{varCat1:}{categorical variable, example: "Genre","Taxonomie", " Nutrition", "Ruche"...}

\item{value:}{value of catVar1 "Lactobacillus" ("Genre"), Taxonomie("Pediococcus pentosaceus"), "Erica cinerea" ("Nutrition"),...}

\item{nc:}{n of clusters (nc=4, default)}

\item{dendrogram:}{dendrogram type ("phylogram", default value), "cladogram", "unrooted", "fan" and "radial"}
}
\value{
output cluster and figure
}
\description{
builds different types of dendrograms for isolates chosen according to varCat1 levels
}
\examples{
   c<-Phyclus(df_Peaks, varCat1="Taxonomie",value="Pediococcus pentosaceus"),
             c<-Phyclus(df_Peaks, varCat1="Nutrition",value="Erica cinerea")
             c<-Phyclus(df_Peaks, varCat1="Taxonomie",value="Pediococcus pentosaceus", dendogram="cladogram")
             source: https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/hclust
                     https://www.rdocumentation.org/packages/ape/versions/5.4-1
}
