% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HClus_dendrograms.R
\name{PhyclusVar}
\alias{PhyclusVar}
\title{PhyclusVar}
\usage{
PhyclusVar(
  df_m,
  meth = "ward.D2",
  d = "euclidean",
  varCat1,
  value,
  varCat2,
  dendrogram = "phylogram",
  ni = 1,
  nf = 10
)
}
\arguments{
\item{ni, nf:}{first and last columns corresponding to categorical variables  (default values, ni=1, nf=10)}

\item{df_m:}{dataframe containing peaks and metadata}

\item{meth:}{hierarchical clustering algorithm ("ward2", default value), other values: "average", "ward.D", "single", "complete", "mcquitty", "median" or "centroid"}

\item{dist:}{distance ("euclidean", default value), "euclidean", "maximum", "manhattan", "canberra", "binary" "minkowski"}

\item{varCat1:}{categorical variable, example: "Taxonomie"    ,"Genre",    "Date.d.analyse"    ,"Origine","Ruche", "Nutrition"    , "Date.de.récolte"    , "Lieu.de.la.ruche"}

\item{value:}{value of catVar1 "Lactobacillus" ("Genre"), Taxonomie("Pediococcus pentosaceus"), "Erica cinerea" ("Nutrition"),...}

\item{varCat2:}{categorical variable}

\item{nc:}{n of clusters (nc=4, default value)}

\item{dendrogram:}{dendrogram type ("phylogram", default value), "cladogram", "unrooted", "fan" and "radial"}
}
\value{
output cophenetic coefficient and figure
 @examples    dft<-PhyclusVar(df_Peaks, varCat1="Genre", value="Lactobacillus", varCat2="Nutrition"),
              dft<-PhyclusVar(df_Peaks, varCat1="Taxonomie", value="Lactobacillus plantarum", varCat2="Ruche")
              dft<-PhyclusVar(df_Peaks, varCat1="Taxonomie", value="Pediococcus pentosaceus", varCat2="Nutrition", dendrogram="cladogram"),
              dft<-PhyclusVar(df_Peaks, varCat1="Taxonomie", value="All", varCat2="Date.de.récolte", dendrogram="fan")
              dft<-PhyclusVar(df_Peaks, varCat1="Taxonomie", value="Lactobacillus plantarum", varCat2="Ruche", dendrogram="fan"),
              dft<-PhyclusVar(df_Peaks, varCat1="Taxonomie", value="Lactobacillus plantarum", varCat2="Ruche", dendrogram="unrooted")
             source: https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/hclust
                     https://www.rdocumentation.org/packages/ape/versions/5.4-1
}
\description{
builds different types of dendrograms for isolates selected and categorized based on varCat2 and varCat1 levels
}
