% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDProcessing.R
\name{PMetaData}
\alias{PMetaData}
\title{PMetaData}
\usage{
PMetaData(
  Dirp,
  fileout = "expMax.csv",
  fu = "max",
  y = 2,
  id = "NRI",
  sc = c(1, 8, 13, 18, 23, 28, 33, 38, 43, 48, 53),
  spNames = c(1, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50)
)
}
\arguments{
\item{Dirp:}{directory (and path) where the csv files are located,}

\item{fileout:}{name of the output file where Maldi identifiers and species names  are found,}

\item{fu:}{this parameter determines whether only one of the scores (y, default value=2) is considered, or instead, the mean, median, minimum or
maximum of the scores corresponding to each of the isolates, "means", "medians", "select", "max" and "min",}

\item{id:}{determine whether unidentified isolates are labeled with "NRI" (default value) or removed ,}

\item{sc:}{columns containing the scores  spNames:columns containing the names of the identified species}
}
\value{
dataframe (and csv file) with three columns: Maldi identifiers, isolate identifications and the corresponding scores
}
\description{
extracts from csv files the taxonomic identifications of the isolates and transforms them according to the values of the scores
}
\examples{
Rn<-PMetaData("Bees project 2019-2020"), Rn<-PMetaData("Bees project 2019-2020", fileout="expMeans.csv", fu="means")
          Rn<-PMetaData("Bees project 2019-2020", fileout="expMedians.csv", fu="medians")
}
