% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusVal.R
\name{PointClusterVal}
\alias{PointClusterVal}
\title{PointClusterVal}
\usage{
PointClusterVal(
  df_m,
  meth = "hclust",
  dist = "euclidean",
  meth2 = "ward.D2",
  varCat1,
  value,
  varCat2
)
}
\arguments{
\item{df_m:}{dataframe containing peaks and metadata}

\item{meth:}{clustering algorithms ("hclust", default value), other values: "kmeans", "pam", "clara", "fanny", "hclust","agnes", "diana"}

\item{dist:}{distances ("euclidean", default value),  "maximum", "manhattan", "canberra", "binary" "minkowski"}

\item{meth2:}{hc methods ("ward.D2", default value), average", "ward.D", "single", "complete", "mcquitty", "median" or "centroid"}

\item{varCat1:}{categorical variable for choosing isolates, examples: "Taxonomie","Genre","Date.d.analyse","Origine","Ruche", "Nutrition", "Date.de.récolte", "Lieu.de.la.ruche"}

\item{value:}{level of catVar1, examples: "Lactobacillus" ("Genre"), Taxonomie("Pediococcus pentosaceus"), "Erica cinerea" ("Nutrition"),...}

\item{varCat2:}{categorical variable for partitioning the set of isolates chosen by using varCat1}
}
\value{
figures and statistics
}
\description{
wrapper of functions for clustering tendency and validation statistics
}
\examples{
   ff<-PointClusterVal(df_Peaks, varCat1="Genre", value="Lactobacillus", varCat2="Nutrition")
       source:https://www.rdocumentation.org/packages/fpc/versions/2.2-8/topics/cluster.stats
              https://www.datanovia.com/en/lessons/cluster-validation-statistics-must-know-methods/
}
