% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SProcessing.R
\name{RowColumnSelector}
\alias{RowColumnSelector}
\title{RowColumnSelector}
\usage{
RowColumnSelector(df_m, varCat1, value, ni = 1, nf = 10)
}
\arguments{
\item{ni, nf:}{first and last columns corresponding to categorical variables  (default values, ni=1, nf=10)}

\item{df_m:}{dataframe containing peaks and metadata}

\item{varCat1:}{selected categorical variable}

\item{value:}{chosen level of varCat1}
}
\description{
selects rows and columns that meet certain conditions determined by varCat1 and  value
}
\examples{
     filtered<-RowColumnSelector(df_Peaks, varCat1="Taxonomie", value="Pediococcus pentosaceus")
}
