% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DClus_MS.R
\name{SMDS}
\alias{SMDS}
\title{SMDS}
\usage{
SMDS(df_m, varCat1, value, varCat2, dist = "euclidean", grah = "lab_mdsGroups")
}
\arguments{
\item{df_m:}{dataframe containing peaks and metadata}

\item{dist:}{distances: "euclidean" (default value), "maximum", "manhattan", "canberra", "binary" or "minkowski"}

\item{varCat1:}{categorical variable for choosing isolates, examples: "Taxonomie"    ,"Genre",    "Date.d.analyse"    ,"Origine","Ruche", "Nutrition"    , "Date.de.récolte"    , "Lieu.de.la.ruche"}

\item{value:}{level of the chosen categorical variable catVar1, examples: "Lactobacillus" ("Genre"), Taxonomie("Pediococcus pentosaceus"), "Erica cinerea" ("Nutrition"),...}

\item{grah:}{graphs: "lab_mdsGroups" (default value), "mdsGroups"}
}
\value{
figures and statistics
}
\description{
function for Multidimensional scaling and  external cluster-based analysis of Maldi_Tof spectra
}
\examples{
SMDS(df_Peaks, varCat1="Genre", value="Lactobacillus",  varCat2="Ruche",dist="euclidean",  grah="lab_mdsGroups")
          SMDS(df_Peaks, varCat1="Genre", value="Lactobacillus" varCat2="Ruche",)
          SMDS(df_Peaks, varCat1="Genre", value="Lactobacillus", varCat2="Nutrition")
          SMDS(df_Peaks, varCat1="Genre", value="All",varCat2="Taxonomie")
}
