% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_PCA_Clus.R
\name{SPCA}
\alias{SPCA}
\title{SPCA}
\usage{
SPCA(
  df_m,
  varCat1,
  value,
  varCat2,
  contDim = TRUE,
  contVar = FALSE,
  contInd = FALSE
)
}
\arguments{
\item{df_m:}{dataframe containing peaks and metadata}

\item{varCat1:}{categorical variable for choosing isolates, examples: "Taxonomie"    ,"Genre",    "Date.d.analyse"    ,"Origine","Ruche", "Nutrition"    , "Date.de.récolte"    , "Lieu.de.la.ruche"}

\item{value:}{level of catVar1, examples: "Lactobacillus" ("Genre"), Taxonomie("Pediococcus pentosaceus"), "Erica cinerea" ("Nutrition"),...}

\item{varCat2:}{categorical variable for partitioning the chosen isolates, examples: Taxonomie,    Genre,    Date.d.analyse,    Origine,    Ruche, Nutrition,    Date.de.récolte, Lieu.de.la.ruche}

\item{contDim:}{graph and statistics for PC contributions (default, contDim=TRUE)}

\item{contVar:}{graph and statistics for variable contributions(default,contVar= FALSE)}

\item{contInd:}{graph and statistics for isolate contributions (default, contInd=TRUE)}
}
\description{
function for Principal Component Analysis of Maldi_Tof spectra with external clusters based on metadata information
}
\examples{
      hg<-SPCA(df_Peaks, varCat1="Genre", value="Lactobacillus", varCat2="Nutrition")
}
