% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusVal.R
\name{VisualOptClusters}
\alias{VisualOptClusters}
\title{VisualOptClusters}
\usage{
VisualOptClusters(
  df_m,
  meth = "hclust",
  meth2 = "ward.D2",
  dist = "euclidean",
  varCat1,
  value,
  nc = 3
)
}
\arguments{
\item{df_m:}{dataframe containing peaks and metadata}

\item{meth:}{clustering algorithms ("hclust", default value), other values: "kmeans", "pam", "clara", "fanny", "hclust","agnes", "diana"}

\item{dist:}{distances ("euclidean", default value),  "maximum", "manhattan", "canberra", "binary" "minkowski"}

\item{meth2:}{hc methods ("ward.D2", default value), average", "ward.D", "single", "complete", "mcquitty", "median" or "centroid"}

\item{varCat1:}{categorical variable for choosing isolates, examples: "Taxonomie"    ,"Genre",    "Date.d.analyse"    ,"Origine","Ruche", "Nutrition"    , "Date.de.récolte"    , "Lieu.de.la.ruche"}

\item{value:}{level of catVar1, examples: "Lactobacillus" ("Genre"), Taxonomie("Pediococcus pentosaceus"), "Erica cinerea" ("Nutrition"),...}

\item{nc:}{number of clusters (nc=3, default value)}
}
\value{
figures and statistics
}
\description{
visualization and statistics for clustering validation
}
\examples{
       s<-VisualOptClusters(df_Peaks, varCat1="Genre", value="Enterococcus"),
                 s<-VisualOptClusters(df_Peaks, meth="pam", varCat1="Genre", value="Lactobacillus", nc=2)
                s<-VisualOptClusters(df_Peaks, meth="hclust", meth2="average", dist="pearson",varCat1="Genre", value="All", nc=5),
                s<-VisualOptClusters(df_Peaks, meth="kmeans", dist="euclidean",varCat1="Genre", value="All", nc=3)
       source:  https://www.rdocumentation.org/packages/factoextra/versions/1.0.7/topics/fviz_cluster
                https://afit-r.github.io/kmeans_clustering
                http://rstudio-pubs-static.s3.amazonaws.com/265632_3ad9e0b981244e15887677f8dffb39a0.html#using-30-different-indices
}
