% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HClus_dendrograms.R
\name{Vizclus}
\alias{Vizclus}
\title{Vizclus}
\usage{
Vizclus(
  df_m,
  meth = "ward.D2",
  dist = "euclidean",
  varCat1,
  value,
  nc = 4,
  graph = "phylogram",
  ni = 1,
  nf = 10
)
}
\arguments{
\item{ni, nf:}{first and last columns corresponding to categorical variables  (default values, ni=1, nf=10)}

\item{df_m:}{dataframe containing peaks and metadata}

\item{meth:}{hierarchical clustering algorithm ("ward2", default value), other values: "average", "ward.D", "single", "complete", "mcquitty", "median" or "centroid"}

\item{dist:}{distance ("euclidean", default value), "euclidean", "maximum", "manhattan", "canberra", "binary" "minkowski"}

\item{varCat1:}{categorical variable for choosing isolates, examples: "Taxonomie"    ,"Genre",    "Date.d.analyse"    ,"Origine","Ruche", "Nutrition"    , "Date.de.récolte"    , "Lieu.de.la.ruche"}

\item{value:}{level of catVar1, examples, "Lactobacillus" ("Genre"), Taxonomie("Pediococcus pentosaceus"), "Erica cinerea" ("Nutrition"),...}

\item{nc:}{n of clusters (nc=4, default value)}

\item{dendrogram:}{dendrogram and factor map}
}
\value{
output figures and statistics
 @examples    Lt<-Vizclus(df_Peaks,varCat1="Taxonomie", value="Pediococcus pentosaceus"),
              Lt<-Vizclus(df_Peaks,varCat1="Ruche", value="VI")
              Lt<-Vizclus(df_Peaks,varCat1="Nutrition", value="Caraway", nc=2),
              Lt<-Vizclus(df_Peaks,varCat1="Genre", value="Lactobacillus", nc=4)
              Lt<-Vizclus(df_Peaks,varCat1="Nutrition", value="All"),
              Lt<-Vizclus(df_Peaks,varCat1="Nutrition", value="Caraway", nc=3, graph="fm")
     source: https://www.rdocumentation.org/packages/stats/versions/3.6.2/topics/hclust
             https://www.rdocumentation.org/packages/factoextra/versions/1.0.7/topics/fviz
             http://www.sthda.com/english/wiki/factoextra-r-package-easy-multivariate-data-analyses-and-elegant-visualization#:~:text=factoextra%20is%20an%20R%20package,exploratory%20multivariate%20data%20analyses%2C%20including%3A&text=Factor%20Analysis%20of%20Mixed%20Data,both%20quantitative%20and%20qualitative%20variables.
             http://www.sthda.com/english/wiki/beautiful-dendrogram-visualizations-in-r-5-must-known-methods-unsupervised-machine-learning
}
\description{
clusters and builds dendrograms for isolates selected according to varCat1 levels
}
