% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DProcessing.R
\name{XmlNCleaning}
\alias{XmlNCleaning}
\title{XmlNCleaning}
\usage{
XmlNCleaning(Dirp, newdir, op = "delete", n = 18, pl = 440518)
}
\arguments{
\item{Dirp:}{folder containing the Maldi_Tof spectrum files,}

\item{newdir:}{folder where the chosen files are copied,}

\item{op:}{"delete"(default value) or "copy",}

\item{n:}{number of lines (n=18, default value),}

\item{number:}{n of characters of line 8 (440518, default value)}
}
\description{
copies all mzXML files with at least n lines to a new folder or delete those with less than n lines, (the mzXML files have 26 lines) this way incomplete or truncated files are removed...the number of characters of line n=18 is also verified
}
\examples{
 XmlNCleaning("newdata")  XmlNCleaning("newdata5", "newdata", op="copy"),XXmlNCleaning("newdata5", "newdata", op="copy", n=26)
           its recommended to run this function with op="delete" after RenameXml,thus cleaning the output directory of the latter function
}
